[WJ] Reorganise low-level memory classes
