+++ go-slang/src/vm/tag.ts
@@ -1,0 +1,12 @@
+export enum Tag {
+  False,
+  True,
+  Number,
+  Nil,
+  Blockframe,
+  Callframe,
+  Closure,
+  Frame,
+  Environment,
+  Builtin,
+}
